Restored 'WM tried to map xx, revert' case.
We do not track window map state very strict, so there are times
when WM may try to map a window that is hidden on vmside and
it should be unmapped xside. ie. when switching desktop.
This is handled by intercepting map message and unmapping the window
back xside. But tray windows are hidden at the begining and are mapped
by the tray manager, so we specialcase them and always map ones.