Implemented qubes-session.
In qubes the Xorg server is started by GUI-Agent.
This commit changes qubes_run_xorg.sh script launched by Agent behavior,
so it starts whole xinitrc session, not just XOrg server.
But we don't use full-blown GNOME or KDE session, instead we have
simplistic qubes-session script which only job (so far) is to parse
/etc/xdg/autostart and launch session applications defined there.
We do this mainly to get the NetworkManager Applet tray icon, but do it
in a generic way that may be usefull in the future.

qubes-session obeys usuall .desktop file keys like:
Hidden, OnlyShowIn and NotShowIn. It uses QUBES or vm_name as session
name string.
So, to enable an application for NetVMs only, add:
OnlyShowIn=NetVM
to the application .desktop file.
And to disable an application in all Qubes VMs add:
NotShowIn=QUBES