gui-daemon: do not call force_on_screen from handle_map()
It is not needed there.
MapEvent does not change window dimensions; so if VM sent evil dimensions
in create or configure_from_vm message, they would have been
corrected once the mesage has arrived.
Apparently, KDE tray temporarily moves the window without resizing so
that it is partially offscreen (which is fine:
we allow window manager do to anything, it is trusted), which
triggered the below check and caused tray malfunction.