vm: pretend that GUI agent is a window manager (#409)
Used to disable broken GTK+ move/resize fallback, which grabs pointer
permanently (so other application in the same VM didn't receve mouse events).
When GTK+ finds that WM supports move/resize - it send a message to WM (which
is ignored) and _do not call_ fallback.

This happens when user clicks on right bottom window corner.