dom0: do not check tray icon position on MSG_MAP
We don't allow tray icons to control its size/position (it is fully managed by
dom0 window manager), so there is no need to check this. Apparently, KDE tray
temporarily moves the window without resizing so that it is partially offscreen
(which is fine: we allow window manager do to anything, it is trusted), which
triggered the below check and caused tray malfunction.