gui-daemon: avoid copying stack state to global shm_cmd
When the client writes less MFN entries than the maximum,
the dom0 side would copy part of its own stack to a global
variable (admittedly, only stuff that had earlier gone
deeper in the stack). The SHM code is still obscure to me
so I'm not sure if the appVM can access this. I think it
cannot, but since there's no reason for copying this, we
just memset instead, so we know there can't be anything
we wouldn't want in there.