pacat-simple-vchan: do not early out when recording is not allowed
I think doing so will not flush the data pulseaudio has, but the
function body will now not write data to the vchan if recording
is not allowed, and we get to call _drop, which should flush the
data. This guards against possibly keeping some leftover audio
that we might not want to have recorded, which could be sent
next time recording is allowed again.