pacat: prevent possible race between rec_allowed and cork ops
The property set/get functions may be called in a different thread
than vchan_rec_callback_data. If unlucky, this could cause the
recording stream to not be corked when it should be. For instance,
in the case where a VM is currently allowed to record, and is not
recording, and the user rescinds permission to record:

- vchan_rec_callback reads a start command
- vchan_rec_callback sets rec_requested to 1
- vchan tests rec_allowed, which is 1
- context switch
- set_property is called with rec_allowed=FALSE
- p->u->rec_allowed is set to FALSE
- pa_stream_cork is called with 1 (stream stops)
- context switch
- pa_stream_cork is called with 0 (stream starts)

This will not actually cause data to be sent to the VM since
the recording callback checks if the recording is allowed
before writing every buffer.