Do not register handler for SIGTERM - exit immediately (#825)
There are nothing to clean (all things rely on open file descriptors,
which are closed by system or have atexit(3) callbacks registered).

The only reason for that handler was to not terminate the process in the
middle of handling some message. But actually if domain is dead, the
process can be waiting on read() or some other syscall and will be
restarted (because of SA_RESTART). Disabling SA_RESTART on SIGTERM have
almost the same effect as terminating the process immediately - the only
error handlers reports the error and either exit the function or
process.