shmoverride: add support for new Xorg binary location
On Fedora >= 21, /usr/bin/Xorg is just a script which calls either
/usr/libexec/Xorg.wrap (which have setuid) or /usr/libexec/Xorg.bin.
/usr/libexec/Xorg.wrap reads /etc/X11/Xwrapper.config to decide whether
drop root privileges or not.
But in any case calling a script will cause shmoverride to be preloaded
to the shell process instead of Xorg binary...
Because shmoverride will unset LD_PRELOAD (to be loaded only once), we
need to call /usr/libexec/Xorg.bin directly, ignoring
/usr/libexec/Xorg.wrap. If someone want to drop Xorg suid, needs to do
it manually on /usr/bin/X-wrapper-qubes.