xside: fix handling resize/move windows on certain window managers
Process both synthetic and real XConfigureEvent. Some window managers
(especially tiling ones) do not send synthetic XConfigureEvent in every
case. This is totally correct behaviour according to ICCCM, so we need
to handle it somehow.
Especially in case of reparenting window manager it can be:
 - window move but not resize: only synthetic XConfigureEvent
 - window resize but not move: only real XConfigureEvent

Some window managers (for example KDE) sends also synthetic
XConfigureEvent, even when not required to do so. Previously we depended
on this behaviour.

Details here: http://tronche.com/gui/x/icccm/sec-4.html#s-4.1.5

To fix this we need to process both synthetic and real XConfigureEvent.
In case of real events, it contains coordinates relative to window
parent. For reparenting window manager it is window decoration created
by window manager - instead of original window parent. Because of that,
we need to translate coordinates. For performance reasons, do it only
when we know that the window is reparented by window manager.