xside: ignore failed XUnmapWindow on non-existing window
It looks like something destroys application windows before GUI daemon
do it. Probably some X requests are reordered and XDestroyWindow got
handled before XUnmapWindow.
Anyway this is rather hard to find actual cause and since unmapped
window in most cases will be destroyed in a moment, simply ignore this
error. If this window was needed and was destroyed unintentionally, gui
daemon will notice it on the next request (this time not ignoring the
error).

Fixes QubesOS/qubes-issues#2085