xside: Fix handling -n flag - don't crach child process
When -n flag is used (do not wait for agent connection), or startup
timeout expires, parent process exits. In such a case, writing
notification for it to a pipe results in SIGPIPE (pipe is closed).
Ignore this signal, so it will not kill the process, instead write
syscall will return EPIPE. Then, we can either handle such error, or
ignore it when needed.

Fixes QubesOS/qubes-issues#2837