gui-daemon: force colorful frame even when VM have not sent window image
Draw colorful frame for override_redirect=True windows just before
mapping the window. To do that, call do_shm_update instructing it to
redraw the whole window. To make it effective, rearrage the case when
shared memory addresses weren't sent yet - to not exit early from the
function, rather skip only XShmPutImage step.

This fixes the case when VM have not sent MSG_MFNDUMP command before
mapping the window - so the composition buffer backing the window content
is not known. And the case when MSG_SHMIMAGE covering window borders
was not sent - for example when window tries to have non-rectangular
shape. Both cases, when happen on a window with override_redirect=True,
result in an undecorated window with a white border, instead of
appropriate VM color.

Reported-by: Eric Larsson <eric.larsson@gmail.com>