pulse: replace dbus control with QubesDB + local socket
Since pacat-simple-vchan not necessary run in dom0 anymore, it needs
some better control mechanism than dbus, to be cross-domain compatible.
Use QubesDB for notification of the audio-input state. But use local
socket + qrexec service (coming in the next commit) for changing the
state, as it gives immediate feedback if it was received by
pacat-simple-vchan.

QubesOS/qubes-issues#4547