Convert qubes.WindowIconUpdater to a socket service
As a result, we have a single service running for all the VMs.
Start it using /etc/xdg/autostart to ensure it's running in the
proper graphical session.

Use (domain, ID) as remote IDs to prevent conflicts between
windows.

Depends on Image.get_from_stream_async, implemented in
QubesOS/qubes-linux-utils#52.