Handle destroyed windows in restack_windows
There is a race condition where the other window gets destroyed
by the time we examine it. To reproduce:

1. create and destroy a lot of windows, for instance by calling
   notify-send repeatedly in dom0:

    while true; do notify-send xxx; done

2. create notifications in VM by also calling notify-send
   repeatedly.

Make sure that the calls in restack_windows() can fail.

For XGetWindowAttributes, we already call it a few times, mostly
on root window. That case can still be fatal, so the error handler
is conditional on resource ID.