Use file descriptor passing for grant tables
This avoids the need for shmoverride when not in dom0, or when not
running qubes that use the legacy mfn-based shared memory mechanism.
The vast majority of the changes are because Xlib does not support file
descriptor passing, so XCB had to be used instead.  XCB is lower-level
and so more code was required.