Delete stale WM_TRANSIENT_FOR hint before mapping a window
If a MSG_MAP is received without WM_TRANSIENT_FOR hint, remove stale
property from the window (if any). According to ICCCM 4.1.2, the window
manager inspect this propert when window is mapped (transition from
Widrawn state), so ensure that stale propery is not there at that time.
In practice, rarely an application reuse the same window with and
without WM_TRANSIENT_FOR hint, but lets handle this corner case
correctly anyway.