Use FD passing for shared memory
This is another attempt at using FD passing for shared memory, at the
cost of still requiring shmoverride.

The reason I am filing it as a PR is to at least give it a
chance to be reviewed, so that the work done on it has a chance of not
being wasted.