shmoverride: avoid keeping state for munmap
By sending a freshly opened Xen file descriptor at each request, the
offset is guaranteed to be 0.  Furthermore, there is no need to call
IOCTL_GNTDEV_UNMAP_GRANT_REF after munmap() ― just calling munmap() is
sufficient.

The only remaining reason to override munmap() is to handle mfn-based
mappings, which may have an offset that is not a multiple of the page
size.  These require correcting the address before the kernel will
accept the munmap() call.  Fortunately, there is no legitimate reason
for the X.org Server to call munmap() with a non-page-aligned address,
so it is safe to unconditionally align the address and pass the aligned
version to the kernel, adjusting the size accordingly.  Therefore, it is
no longer necessary for shmoverride to maintain any state whatsoever
about the current mappings, simplifying the code enormously.  No mutex
locking is needed, and no list of mapped pages needs to be kept.  The
result is that the code is simpler and less prone to bugs.