Ignore override-redirect in MSG_CONFIGURE
There is no legitimate case for setting (or unsetting) override-redirect
when moving the window. This especially may be harmful if the window is
already mapped - the window manager may not notice the change and not
start/stop managing the window (including drawing borders and
positioning the window appropriately).
Since override_redirect flag is included in MSG_MAP anyway, simply
ignore it in the MSG_CONFIGURE.

Reported-by: Demi Marie Obenour <demi@invisiblethingslab.com>