Prevent changing override-redirect of a mapped window
Similar to the previous commit, add few more safety checks to avoid
changing override-redirect if the window is already mapped. This
includes sending MSG_MAP (with different parameters) for already mapped
window.
As a safety check, refuse change in set_override_redirect() that is
responsible for setting the actual vm_window->override_redirect.

Reported-by: Demi Marie Obenour <demi@invisiblethingslab.com>