Set window background color to white
This provides some default window content filling, if the VM fails to
update it with the actual content. In that case, window (or part of it),
could be seen as a stale screen content of that place, with just forced
colorful border.
This specifically affects some tray icons (sdwdate-gui or nm-applet
if the network manager daemon is not running), but also some
applications that try to be transparent (like, a green frame window
drawn by zoom when sharing the screen). This change makes such windows
solid white, instead of potentially confusing stale content of the
screen.

QubesOS/qubes-issues#6856