Bidirectional GUI protocol negotiation
This implements GUI protocol v1.4, which includes bidirectional protocol
negotiation.  This means that not only does the agent tell the daemon
what protocol version it supports, the daemon also tells the agent what
protocol version it decided to actually use.  This may be less than the
version supported by the agent.  As a result, a GUI daemon including
this patch may be used with any version of the GUI agent supporting
protocol major version 1, even ones that support higher minor versions
than the daemon does.