Release MFNs for unmapped windows as well
This commit ensures that shared memory grant references belonging to
unmapped windows are released right after the call to XUnmapWindow.

The issue described in QubesOS/qubes-issues#7539 indicates that Xen
grant references are not released for certain GUI windows. Modifying the
in-VM GUI agent to dump Xen grant entry numbers and correlating these
entry numbers with the kernel's logs indicated that these windows were
unmapped but not destroyed.

One reliable manner to reproduce this issue on an up-to-date Qubes OS
v4.1 system with an AppVM based on an Fedora 36 template is starting the
xfce4-terminal program, followed by clicking on the "Help->About" menu
item. If the user waits for more than 60 seconds while keeping the
"About" window open, to exhaust the kernel's pending grant entry timer,
the AppVM's kernel eventually outputs messages such as the following, as
observed via "dmesg -w":

  xen:grant_table: g.e. 0x1234 still pending

(In the author's tests, the 60 second time-out was modified to be a
kernel parameter to trigger the kernel's messages faster.)

This issue may also be related to the following messages emitted by the
GUI daemon (running in dom0) when the "About" window and the parent
xfce4-terminal window are finally closed, which appear to confirm that
the transient "About" window is unmapped but not destroyed right away by
the GUI toolkit running in the AppVM:

  Window 0x6000001 is still set as transient_for for a 0x6000002 \
  window, but VM tried to destroy it

This commit aims to resolve the potential resource exhaustion issue
caused by lingering Xen shared memory grant references that trigger the
aforementioned kernel warnings.

Please note that this commit was built and tested on the release4.1
branch only.

Fixes QubesOS/qubes-issues#7539

Signed-off-by: M. Vefa Bicakci <m.v.b@runbox.com>