Strictly validate message lengths
Previously, message lengths were not strictly validated.  This prevents
relying on the untrusted_len field for message framing.  If some code
does rely on this field, but the value in this field does not match the
number of bytes the GUI daemon actually reads, chaos will ensue.

To catch this problem before it can create havoc, validate that the
untrusted_len field of incoming messages is correct.  In protocol
versions 1.6 and later, length mismatches are fatal.  In earlier
protocol versions, length mismatches are not fatal, as there might be
existing agents that do not set untrusted_len correctly.