Do not use VERIFY with partially-read messages
If the VERIFY returns, this will leave the rest of the message in the
buffer, which in turn will be interpreted as GUI messages.  The results
will be undefined and possibly harmful.  Instead, consider this to be a
fatal error and exit.  The sanitization of message types was just
dropped, as none of the code turns out to rely on it.  Message width
and height checking was simplified to only check that the values are not
too large, as the relevant struct fields are unsigned and so cannot be
negative.