Always block on vchan I/O by default
This significantly improves audio quality compared to nonblocking mode.
I believe this is because the PipeWire agent will usually be able to
free up space in the capture vchan quickly enough to avoid an overrun.

This has not been tested directly, but it is equivalent to always
passing -b as a command-line argument, which I did test.

Fixes: QubesOS/qubes-issues#8887