qubes-guid: add window_background_color setting
This configuration setting allows changing the background pixel for
local windows that display content of VM windows. Generally the color is
only visible rarely and briefly. Previous to this commit the color was
hardcoded to white which is suboptimal for users running dark desktop
styling.

This is the downstream translation of these VM features:

- gui-window-background-color
- gui-default-window-background-color

Testing process:

  [user@dom0 ~]$ qvm-features dom0 gui-default-window-background-color <color>
  [user@dom0 ~]$ qvm-run --dispvm=default-dvm --service qubes.StartApp+debian-xterm
  # (Then play with the terminal's window to expose the color)

Related to: QubesOS/qubes-issues#9304