Add Xwayland wrapper similar to the one for Xorg
Add a wrapper that sets LD_PRELOAD=.../shmoverride.so. Since suid
binaries are not involved with Xwayland, it can be a simple shell
script.

Place the wrapper as Xwayland in /usr/libexec/qubes/wrappers. Processes
that are supposed to use it need to have /usr/libexec/qubes/wrappers
prepended to PATH.

QubesOS/qubes-issues#8515