shmoverride: do not remove the shmid file if taking the lock fails
If shmoverride init fails to take the lock, it likely means it is used
by currently running process - do _not_ remove the file in such a case.
This may happen if LD_PRELOAD leaks to child processes (which seems to
happend on Xwayland when shmoverride is linked with -z initfirst).
But also, if X server is started several times for some reason (for
example user calls it manually), it will abort, but not early enough to
skip shmoverride init.

QubesOS/qubes-issues#8515