shmoverride: allow applying special behavior only to specific programs
This allows setting LD_PRELOAD earlier for some parent process, and have
it actually applied to selected child process.
This is especially helpful for Xwayland that is started by the wayland
compositor and rarely has an option to change binary path (contrary to
display managers starting Xorg). This feature allows setting LD_PRELOAD
on the compositor itself, and have it affected only Xwayland process.

Note, this feature does not work if shmoverride is linked with -z
initfirst, as getenv()/unsetenv() are not working in that case.

QubesOS/qubes-issues#8515