audio: fix reacting on mic enable/disable requests
First of all, after getting an event of QubesDB FD, actually call
qdb_read_watch() so the event is cleared and the next one can be
received.

But that's not enough. libqubesdb's queueing of watches doesn't handle
well using the same connection for both watches and commands, and using
non-blocking mode (monitoring qdb_watch_fd(), instead of calling
qdb_read_watch() in a loop). While this issue wants fixing on the
libqubesdb side, for now use separate connection to workaround the
issue.

Fixes QubesOS/qubes-issues#9619