debian: fix installing qubes-guid as suid
This is how it works on Fedora, so mirror the behavior in Debian package
too.
Technically it might be possible to avoid the suid, by adjusting at
least log dir permissions, and probably few more. But that's separate
change, unbreak Debian for now.

Use postinst script instead of packaging with desired permissions
already, because the qubes group doesn't exist at the package build
time.

QubesOS/qubes-issues#8841